calculate skill rankings by only one thread at once

+++ ysg-manager-server/src/main/java/com/lukas99/ysgmanager/domain/SkillService.java
@@ -2,7 +2,9 @@ package com.lukas99.ysgmanager.domain;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.concurrent.Semaphore;
 import javax.persistence.EntityManager;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -12,6 +14,7 @@ import org.springframework.transaction.annotation.Transactional;
  */
 @Service
 @Transactional
+@Slf4j
 public class SkillService {
 
   private final SkillRepository skillRepository;
@@ -92,19 +95,34 @@ public class SkillService {
   }
 
   /**
+   * To ensure that skill ranking calculation is done by only one thread at once.
+   */
+  private final Semaphore skillCalculationSemaphore = new Semaphore(1);
+
+  /**
    * Calculates the skill rankings and the tournament skill rankings for the given tournament.
    *
    * @param tournament The tournament for which the rankings should be calculated.
    */
   public void calculateSkillRankings(Tournament tournament) {
-    skillRankingService.deleteAll(tournament);
-    skillTournamentRankingService.deleteAll(tournament);
-    // without flush, removal will not be persisted and adding new rankings
-    // will cause unique constraint violation exceptions
-    entityManager.flush();
+    if (skillCalculationSemaphore.tryAcquire()) {
+      log.info("Start skill ranking calculation");
 
-    skillRankingCalculator.calculateRankings(tournament);
-    skillTournamentRankingCalculator.calculateTournamentRankings(tournament);
+      skillRankingService.deleteAll(tournament);
+      skillTournamentRankingService.deleteAll(tournament);
+      // without flush, removal will not be persisted and adding new rankings
+      // will cause unique constraint violation exceptions
+      entityManager.flush();
+
+      skillRankingCalculator.calculateRankings(tournament);
+      skillTournamentRankingCalculator.calculateTournamentRankings(tournament);
+
+      skillCalculationSemaphore.release();
+      log.info("Finished skill ranking calculation");
+    } else {
+      // do nothing, skill calculation can only be done by
+      log.info("Skill ranking calculation is already running. Request is ignored.");
+    }
   }
 
 }
